Refactor exception handling to Service layer to throw exceptions to Global Exception Handler

+++ activity2/spring-maven/src/main/java/sg/edu/ntu/springmaven/service/ProductService.java
@@ -6,6 +6,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import sg.edu.ntu.springmaven.POJO.Product;
+import sg.edu.ntu.springmaven.exception.ProductNotCreatedException;
+import sg.edu.ntu.springmaven.exception.ProductNotFoundException;
 import sg.edu.ntu.springmaven.repository.ProductRepository;
 
 @Service
@@ -20,14 +22,25 @@ public class ProductService {
     }
 
     public List<Product> getAllProducts() {
+        if (productRepository.getAllProducts().isEmpty()) {
+            throw new ProductNotFoundException("Products not found");
+        }
         return productRepository.getAllProducts();
     }
 
     public Product getProduct(String id) {
+        Product product = productRepository.getProduct(id);
+        if (product == null) {
+            throw new ProductNotFoundException("Product not found for ID: " + id);
+        }
         return productRepository.getProduct(id);
     }
 
     public Product addProduct(Product product) {
+        Product newProduct = productRepository.addProduct(product);
+        if (newProduct == null) {
+            throw new ProductNotCreatedException("Could not create Product, check Product fields and try again");
+        }
         return productRepository.addProduct(product);
     }
 
@@ -36,6 +49,10 @@ public class ProductService {
     }
 
     public Product deleteProduct(String id) {
+        Product product = productRepository.deleteProduct(id);
+        if (product == null) {
+            throw new ProductNotFoundException("Product not found for ID: " + id);
+        }
         return productRepository.deleteProduct(id);
     }
